refactor: creates queries for admin purpose vs public
